refactor(downloads): extract IPC registration to lib/ipc/registerDownloadHandlers.ts
Pure relocation. Mirrors the existing register*Handlers.ts files in
src/main/lib/ipc/. comfyDownloadManager.ts no longer reaches for
ipcMain / shell directly.

Amp-Thread-ID: https://ampcode.com/threads/T-019e2311-6aeb-7349-9fd9-560288cfbe38
