+++ src/main/lib/comfyDownloadManager.ts
@@ -1,4 +1,4 @@
-import { app, BrowserWindow, dialog, ipcMain, shell } from 'electron'
+import { app, BrowserWindow, dialog } from 'electron'
 import { EventEmitter } from 'events'
 import fs from 'fs'
 import path from 'path'
@@ -840,44 +840,4 @@ export function _test_setSeededTrayState(snapshot: DownloadsTrayState): void {
   downloadEvents.emit('tray-state-changed')
 }
 
-// ---- IPC registration ----
 
-export function registerDownloadIpc(): void {
-  ipcMain.handle(
-    'desktop2-download-model',
-    (event, { url, filename, directory }: { url: string; filename: string; directory: string }) => {
-      const win = BrowserWindow.fromWebContents(event.sender)
-      if (!win) return false
-      return startModelDownload(win, url, filename, directory, event.sender)
-    },
-  )
-
-  ipcMain.handle('model-download-pause', (_event, { url }: { url: string }) =>
-    pauseModelDownload(url),
-  )
-
-  ipcMain.handle('model-download-resume', (_event, { url }: { url: string }) =>
-    resumeModelDownload(url),
-  )
-
-  ipcMain.handle('model-download-cancel', (_event, { url }: { url: string }) =>
-    cancelModelDownload(url),
-  )
-
-  ipcMain.handle('model-download-dismiss', (_event, { url }: { url: string }) =>
-    dismissRecentDownload(url),
-  )
-
-  ipcMain.handle('model-download-clear-finished', () => clearFinishedDownloads())
-
-  // Seed the renderer-side store with active entries AND the recent
-  // terminal buffer so the Settings tab + popup history are non-empty
-  // on first paint after a window opens mid-flow.
-  ipcMain.handle('model-download-list', () => getAllDownloads())
-
-  ipcMain.handle('show-download-in-folder', (_event, { savePath }: { savePath: string }) => {
-    if (typeof savePath === 'string' && savePath) {
-      shell.showItemInFolder(path.resolve(savePath))
-    }
-  })
-}
